+++ sib-it/src/main/java/com/fxclub/account/service/impl/AccountServiceImpl.java
@@ -10,6 +10,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
+
 @Slf4j
 @Service
 public class AccountServiceImpl implements AccountService {
@@ -36,10 +38,10 @@ public class AccountServiceImpl implements AccountService {
 
     @Override
     @Transactional
-    public void deposit(Long id, double amount) {
+    public void deposit(Long id, BigDecimal amount) {
         Account account = repository.getOne(id);
         if (account != null) {
-            account.deposit(amount);
+            account.deposit(amount.doubleValue());
             repository.save(account);
         }
     }
